feat(converter): use float as progress meter

+++ pkg/media/converter_models.go
@@ -33,7 +33,7 @@ type ConvertProgress struct {
 	CurrentBitrate  string
 	Progress        float64
 	Speed           string
-	FPS             int
+	FPS             float64
 	File            files.Filer
 }
